Add opa-fmt rule (#62)
As this was the first linter rule in Go, it took quite some
plumbing to get right. I'd expect (and hope!) future Go rules
will be less demanding.

I've also refactored the input passed to the various linting
functions so that it now includes the raw data from the original
file as well. This was needed for comparison for `opa fmt`, but
will also be useful later for linter rules that need to work on
data not found in the AST.

The .rego file changes here can largely be ignored, as the
only change is running the formatter on them.

Resolves #40

Signed-off-by: Anders Eknert <anders@styra.com>