Add parsing of ignore directives (#67)
It is now possible to add an "ignore directive" above any
line of Rego, to have Regal ignore the next line given that the
rule(s) named in the directive match that of the linter being
executed. Example:

```rego
\# regal ignore:prefer-snake-case,use-assignment-operator
default fooBar = "baz"
```

Note that this only works for the next _line_, and does not take
into account concepts like packages, functions or rules. But this
should be good enough for most use cases, I think.

Resolves #39

Signed-off-by: Anders Eknert <anders@styra.com>