Add command for generating markdown table from rule annotations
And generate a table for our README :)

NOTE: This should be integrated in the build, so that whenever a
built-in is added or updated, a new table is generated. I plan to
look into build-related improvements later, so doing this manually
for now.

Additionally, configure `gci` linter to group imports.

Fixes #36

Signed-off-by: Anders Eknert <anders@styra.com>