refactor: one rule, one file
Having all rules lumped into a single file only per
category got to be a bit messy eventually. This change
uses one file per rule, making it easier to both find and
to work with rules in isolation — hopefully making it easier
to contribute new ones too.

There is still a few things I would like to fix, like having
unit tests actually be unit tests, and not as it currently stands
that they test an input against *all* rules in the category...
but this change is big enough as it is, so I'm deferring that to
a future PR.

Fixes #144

Signed-off-by: Anders Eknert <anders@styra.com>