Enhanced type checking using Regal AST schema
Currently enabled for the `regal test` command, which is the most
likely entrypoint for rule authors, whether for built-in Regal rules
or custom ones.

While it would be useful to provide this also for `regal lint` — we'll
probably want to do this together with the strictness rules, as we
currently don't require compilation, and it's not clear that we'll want
to unless a rule is configured that needs this, like most of the strictness
rules would.

Additionally, the schema is provided when running the Rego tests via
`go test ./...`, which should help catch errors we make ourselves.

It would be nice if custom rule authors didn't have to include the
schema in the metadata annotation, but I've yet to figure out the best
way of doing that, as we can't bundle a custom policy and also have them
read from disk without the bundle "owning" that path. I suppose we could
just parse a policy from a string, including a subpackage scoped
custom.regal.rules package annotation... but I'll need to think more
about this.

Fixes #52

Signed-off-by: Anders Eknert <anders@styra.com>