Rule: prefer-some-in-iteration
Building on top of previous work to find vars and to determine the
"scope" of a variable, this commit adds a new rule in the style
category to prefer `some x in y` over `x := y[_]`.

There are also a few fixups here related to the new rule, and some
that were forgotten about in the previous rule added.

Note that there are some situations where a variable may thought of
as being in scope while not actually being in scope — such as when
iteration happens after a comprehension, as we have no way of keeping
the nesting level "state" currently. This should not lead to false
positives but could lead to a few cases being missed, which I think
is OK for the time being. Will need to investigate better methods
for determining scope, and likely outside of Rego.

Fixes #243

Signed-off-by: Anders Eknert <anders@styra.com>