Rule: use-some-for-output-vars
This commit adds AST functions to find vars in scope of a location,
which means scanning "up" in the scope for e.g. rules or other vars
to check if they are present. This allows us to see if a possible
"output var" is defined using some, or if it is an "input" var that's
assigned elsewhere in the policy.

Fixes #250

Signed-off-by: Anders Eknert <anders@styra.com>