Allow not providing level in user configuration
Previously, this would lead to the rule level getting set to "" (empty string),
which "worked" as it would still print the violation, but since it wasn't set
to `error` it would not change the exit code of the lint command.

If the level is not provided in user configuration now, we simply inherit the
value from the provided configuration.

Fixes #285

Signed-off-by: Anders Eknert <anders@styra.com>