Check for presence of binary before running e2e tests
Rather than causing a panic, exit gracefully with a message
in case the binary doesn't exist.

I entertained the idea of building the binary if needed, but
since these tests run in parallel, that would be problematic.

Tried using the [TestMain](https://pkg.go.dev/testing#hdr-Main)
function to check only once, but there seems to be no way to
get print output from that function unless `-v` is used, as
none of the t.Log functions are available there.

Fixes #294

Signed-off-by: Anders Eknert <anders@styra.com>