build/do.rq: build binary for tasks list if needed
`build` is not `build(fresh)`, and will only rebuild the ./regal binary
if passed `true`. If passed `false`, it will not rebuild the binary if
some binary is present in the given location.

`build(false)` is thus to be used when any regal version will do, like
for printing tasks.

Signed-off-by: Stephan Renatus <stephan@styra.com>