Rule: `prefer-package-imports`
This builds upon the design laid out in #282,
and on @sesponda's excellent work on that. My
approach for testing this was to implement an
actual "real world" linter rule from the backlog,
and `prefer-package-imports` seemed like a good
choice. The rule aggregates package names (or paths)
from each input file, along with any import paths.

The report rule then checks each import collected
against the list of package paths to determine if
the import matches a package. If not, the rule will
consider that a violation.

As is often the case when real world meets the original
design, "some" modifications proved to be necessary.

The collection of aggregate data is now run alongside
the usual linter rule evaluation, as we conveniently
may use the same input files for aggregating data.
Once that's done, we now send another eval call using
only the data aggregated as input. Each aggregate rule
is now provided only the data aggregated by _that rule_,
so there's no need for policy authors to traverse through
all of the aggregates, but one may simply refer to the
`input.aggregate` array for the same set of entries as
one collected previously.

I've also had to add an alternative schema for the input
when these rules are evaluated (currently only enforced
by `regal test`). Perhaps `anyOf` would be a better
option than two separate schemas, but I'll explore that
at a later point.

Fixes #326
Fixes #343

Signed-off-by: Anders Eknert <anders@styra.com>