Rule: `equals-pattern-matching`
Prefer to use pattern matching for equality in function args.

Currently limited to simple rules where the equality check is
the only condition in the rule head or rule body (and this
limitation is described in the rule docs). I'm hoping to improve
this in the future, but as I think there's value in this already,
so let's ship and iterate.

Fixes #397

Signed-off-by: Anders Eknert <anders@styra.com>