Optimization: less use of walk (#455)
* Refactor: collect rules to run only once

For the provided rules, we don't need to check this over
and over. Not making a dent on eval time, but somewhat nicer
looking code.

Signed-off-by: Anders Eknert <anders@styra.com>

* Optimization: less use of walk

The rules that walk refs may benefit from having the AST tree
walked *once* with refs collected, and then simply iterate over
that collection. Also, a few  micro-optimizations in the hotspots
reported by `--profile`.

Eval time reduced by 4-5%, so nothing dramatic, but still good.

Signed-off-by: Anders Eknert <anders@styra.com>

---------

Signed-off-by: Anders Eknert <anders@styra.com>