Use `import rego.v1`
Note that for now, this change only affects what
we use internally in Regal, and not what we enforce.

This will be done too in due time, but given that
`rego_v1_import` isn't a capability in any released
version of OPA yet, we should hold off on recommending
`import rego.v1` until there is, and the feature works
without known bugs, which will be from the next release.

Signed-off-by: Anders Eknert <anders@styra.com>