Concurrent parse in `InputFromPaths`
Good news is this reduced the total time to parse,
as expected. The bad news is that this wasn't the
metric I thought it was, and the real killer is
`timer_rego_input_parse_ns`, which we can't fix as
easily.

Oh well, a small win is still a win.

Fixes #462

Signed-off-by: Anders Eknert <anders@styra.com>