Rule: `duplicate-rule`
As the name implies, flag duplicate rules when found. See the docs included
in the PR for more details, including current limitations.

This rule should be an aggregate rule, but it was hairy enough to pull off
with a single file scope, so I decided to hold off on that for now. This
provides value already, so let's ship it and consider looking into making
this aggregate later.

Fixes #427

Signed-off-by: Anders Eknert <anders@styra.com>