Address `external-reference` missing cases
This took quite some effort as the first iteration of this rule
was rather basic. While this makes it a bit annoying with its
pedantry, at least it's now **correct**. One benefit of this is
that `ast.find_vars` now also takes function arg return value
bindings into account, which may have positive impact on other
rules as well.

Fixes #191

Signed-off-by: Anders Eknert <anders@styra.com>