Publish binaries from build workflow
No Windows binaries as of yet since `build/do.rq pull_request` silently
fails, or does nothing, on Windows. The big discovery here however — we
never actually built on Windows in these worflows before, or even on MacOS!
This since the `runs-on` attribute didn't make use of the matrix OS value, but
was instead hard-coded to use `ubuntu-latest` 🙃

We'll have to get back to the Windows build at some point, but for now we'll
settle on only building on that platform for releases via goreleaser, as we
unknowingly seem to have done up until this point anyway.

Fixes #608

Signed-off-by: Anders Eknert <anders@styra.com>