lsp: `textDocument/hover` implementation (#617)
This PR adds support for the `textDocument/hover` feature of the
language server protocol. Currently, only the built-in functions
of OPA are supported, which likely satisfies the most common
requirement. In the future, we could extend this to cover also
user defined functions or rules, where annotations are used to
provide data for the hover response.

Signed-off-by: Anders Eknert <anders@styra.com>