lsp: add inlay hint support
This PR adds support for inlay hints of function arguments.
Similar to the hover feature, only built-in functions are
supported at this point.

Also did some minor refactoring in the hover function to
have it use the same method for collecting built-in functions
from a module as the inlay hint system does.

Signed-off-by: Anders Eknert <anders@styra.com>