Refactor LSP logging
While working on https://github.com/StyraInc/regal/pull/630, I needed
to be able to see requests, responses and notifications to the server
from the client and from the server to the client.

I learned that there is functionality in the jsonrpc2 lib that we use to
do this nicely. This PR implements a new logging configuration to make
use of this using an adapted version of the LogMessage function that is
found in the library and using the OnRecv and OnSend functions.

Because of this, I have removed the logging responsibility from the
server and moved it to the connection. The server now should log errors
only and that should be the default mode of operation.

Enabling verbose logging will show you all messages being sent on the
connection. ExcludeMethods and IncludeMethods can be used to filter the
stream of messages when you are interested in monitoring or ignoring
certain flows of messages. I know that this will be useful for me and I
hope that this is a sensible enough way to do this overall.

Signed-off-by: Charlie Egan <charlie@styra.com>