lsp: Accept but ignore cancelled requests
I have not been able to work out why these are sent but they don't seem
to impact the running functionality of the server and supporting them
seems to be optional.

https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#dollarRequests

For now, this updates the server to do nothing when receiving such a
request. This avoids some log lines that we don't do anything about.

I would like to come back to this though as it seems like there's more
here that we might be missing. So it has been labelled as TODO.

Signed-off-by: Charlie Egan <charlie@styra.com>