lsp: Increase test buffered channel size
TestLanguageServerMultipleFiles is sometimes locking up as the handler
can't save new messages as the channel is waiting to be read and the
part of the test that is waiting to read those particular messages is
later and waiting for another file first.

This makes the buffered channels larger to allow more messages to be
received for each file so that we can no block and check the messages we
want in order as before.

Signed-off-by: Charlie Egan <charlie@styra.com>