lsp: handle `textDocument/documentSymbol` (#668)
More weekend fun! This provides a basic implementation of `textDocument/documentSymbol`.
I say basic because we currently scan for nested symbols, but only package, rules and
functions, with some nice touches like providing details on what kind of rule it is, and
the type it returns, and so on. I think this covers what most actually need, but we can
certainly refine this later and add things like variables inside of rule bodies, and so on.

Signed-off-by: Anders Eknert <anders@styra.com>