lsp: More graceful shutdown
We have been erroring on shutdown requests as they have no params set.
This PR changes that and makes sure the server waits for the exit
message before closing the connection.

The client shows:

```
[Error - 18:21:10] Server process exited with code 0.
```

This is an odd one based on the code here:

https://sourcegraph.com/npm/vscode-languageclient@492942af8005ce98cbd47721d4a7790a0310d4e3/-/blob/lib/node/main.js?L474:37-474:68

It would look like we exit ok here, but that seems to be perhaps parsed
as "0" the string... One for another day.

Signed-off-by: Charlie Egan <charlie@styra.com>