Avoid panic when searching for `.regal/config.yaml` (#684)
As reported in #682, the server panics when the dir for whatever reason
is "". I don't yet know how to trigger this, and the result of this fix
is likely that the config file won't be found when this happens. But as
a temporary measure I prefer to have Regal run with default configuration
compared to crashing.

Signed-off-by: Anders Eknert <anders@styra.com>