lsp: Fix issue with parse error inlay hints
This change updates the server to attempt a partial parse instead when
there are parse errors to still, if possible, show the hints for
sections of the file before the first parse error.

This is a best effort adjustment, and if the parse still fails, all
inlay hints are cleared until one or all of the parse errors are
addressed.

Signed-off-by: Charlie Egan <charlie@styra.com>