Send warning to client on CRLF line endings
In case either `opa-fmt` or `use-rego-v1` is enabled, we'll send back
a warning to the client whenever they save a document containing CRLF
line endings. This to avoid confusion when the user tries to fix formatting
using `opa fmt`, but the editor adding back CRLF line endings after.

Also disable the nestif linter, as it's annoying.

Fixes #693

Signed-off-by: Anders Eknert <anders@styra.com>