Check Regal version at start-up
This adds functionality to the regal LS and lint commands what will
check the current version against the latest release in GH.

In order to keep the number of requests down, the version is cached to
disk in the .regal dir. The task runs in another goroutine to keep
things fast too.

Config has been updated with a features key, and
REGAL_FEATURES_REMOTE_CHECK_VERSION can also be set to disable to
functionality.

Sharing for review, if we think it looks good, I'll add some docs and we
can get this in.

Signed-off-by: Charlie Egan <charlie@styra.com>