Don't call completion providers inside of comments (#831)
It seems like this is handled client side by some editors, but Zed
triggered completion requests inside of comments. While we could have
each provider handle this, I think an early exit in the manager already
is a better solution, as there's unlikely to be any provider that wants
to do suggestions in a comment. If that changes in the future, we'll
handle it then.

Fixes #826

Signed-off-by: Anders Eknert <anders@styra.com>