Add snippets provider
Snippets are special inserts that allow users to enter values and press tab
to move to the next location.

https://code.visualstudio.com/docs/editor/userdefinedsnippets

Add snippets for:
- `some v in coll`,
- `some k, v in coll`
- `every v in coll`
- `every k, v in coll`

Got interested in trying this after learniung that Zed supports snippets in
the latest release. I've tried and confirmed this works in both VS Code and
in Zed.

Signed-off-by: Anders Eknert <anders@styra.com>