Begin to Include end location where applicable
This PR attempts to establish end location as an optional attribute of location
data returned in violations, and updates ~10 rules in the bugs category to use
this convention.

If no end location is provided, the default is as it used to be — end location
will be set to the end of the line. For some rules, this is the right thing to
do anyway, and therefore, not all rules will need to be updated..

That obviously doesn't "fix" #654, but it's a start, and as far as I've seen so
far, works quite well. Importantly this change is non-breaking method for API
users! Ideally, we'd put the current "row" and "col" into a "start" object too,
but that *would* be a breaking change, and one that would impact all API users,
so let's consider that for our first stable release instead.

NOTE that this does not change anything in the output of `regal lint` using the
default "pretty"-reporter, but is only included in the more detailed JSON and
SARIF formats. First and foremost this change helps LSP clients display
diagnostics more accurately, and that is really where this is most useful.

Signed-off-by: Anders Eknert <anders@styra.com>