Fix bug causing `prefer-some-in-iteration` not to be reported
This was due to an unfortunate combination of trying to be lenient
about `coll[_]` used in the position of a function arg (as that's not
*directly* replaceable by `some .. in` **but** not excluding operators
like `:=` from functions (where it _should_ be replaced by `some .. in`)
**and** many tests for this rule missing to provide the built-in functions
as a mocked dependency.

Will need to think some about how to best avoid this to happen again.

Also:
- Replace `ast.policy` in tests with `ast.with_rego_v1`
- Fix docs that used `=` for assignment rather than `:=`

Fixes #901

Signed-off-by: Anders Eknert <anders@styra.com>