lsp: No error when missing keywords for hover
This means that the keywords for the file have not been computed yet and
should be shortly.

There is unlikely much that this message would help with so I have
removed it.

The error when processing hover content updates should be enough to
establish that the keywords are not being computed, if that ever were to
be the case.

Signed-off-by: Charlie Egan <charlie@styra.com>