Fix missing locals completions
This was caused by us no longer passing JSONOPtions to the parser
that parsed workspace files, and so when passed to the store, the
locations went missing in the roundtrip the store would do before
persisting the data.

The store is now created with the roundtrip option disabled, and
we'll make sure ourselves that the data is serialized correctly,
and in the case of modules — serialized to RoAST. This should
mean less data stored in `data.workspace.parsed` and lower latency
for policies that rely on that store, which is nice!

Also:
- bump golangci-lint and address some new issues

Fixes #1023

Signed-off-by: Anders Eknert <anders@styra.com>