100% test coverage
Like with previous PR aiming to increase our test coverage (#1058),
doing so this time revealed a few places where code was unused, or
could be rewritten to be better optimized, removed, or made easier
to read.

One find that is yet to be addressed here is how we copy the whole
`ast.found.refs` object into `ast.all_refs`, which surely comes with
a cost that seems unnecessary, given we can use `ast.found.refs`
directly instead. But since there are a few consumers, and they are
not **exactly** identical (`all_refs` contains import refs too),
I'll follow up on that in a later PR.

I'm also adding the codecov badge to the README to make it easy to
see when our coverage drops. But I'm not as convinced we should
fail PRs when coverage isn't 100% as I was before, as it might make
for a worse contributor experience.

Fixes #918

Signed-off-by: Anders Eknert <anders@styra.com>