Add source action to explore compiler stages (#1096)
This PR integrates @srenatus excellent
[opa-explorer](https://github.com/srenatus/opa-explorer/)
with Regal, where we now start a web server along with the
language server to host the explorer.

Users will now see a "Source Action" in the context menu
of any given Rego file, and when clicked will bring them
to the explorer view for that file.

Since a command for opening web pages client side is AFAIK
only available in VS Code, this feature is sadly limited
to that editor currently. Perhaps we can figure out some
way to have the link show in Zed, similar to what we did
for linter docs?

For now, we're only loading a single file into the explorer
for compilation. We could perhaps consider loading all
modules in the workspace too, as otherwise compilation will
always fail on e.g. unknown functions from outside the policy.

Signed-off-by: Anders Eknert <anders@styra.com>