lsp: Update contents before diagnosticRequestFile
One of the main causes for the 'empty module' error stemmed from the
sending of diagnosticRequestFile events with no content.

One solution would have been to fetch the content when processing
events on the diagnosticRequestFile channel, however it felt like the
publisher had more knowledge of how to do this correctly and
processing diagnosticRequestFile was getting unwieldy.

I have removed the content from diagnosticRequestFile, and these events
will just update the parse and send diagnostics. This should be less
surprising.

Signed-off-by: Charlie Egan <charlie@styra.com>