lsp: Address test concurrency issues
This PR is intended to address two issues with the test suite that
happen when running tests concurrently:

TestLanguageServerSingleFile fails sometimes as the server logs an error
when the context has been cancelled, and it's a panic when logging to t,
after the test is over.

https://github.com/StyraInc/regal/actions/runs/10875806798/job/30174833322

TestGetInlayHintsAstTerms will fail when TestLanguageServerSingleFile is
running sometimes as TestLanguageServerSingleFile updates the contents
of the map while TestGetInlayHintsAstTerms is iterating over it.

https://github.com/StyraInc/regal/actions/runs/10875806798/job/30184826633

Signed-off-by: Charlie Egan <charlie@styra.com>