fixer: add --on-conflict flag to support renaming
This PR adds a new flag to the fix command, `--on-conflict`. This flag
can be used to specify the behavior of the fix command when a file's new
name conflicts with an existing file or another file's fixed name.

Setting --on-conflict to "rename" will cause the fix command to rename
the file to a different name at the target location. The new name will
be generated by appending a number to the end of the file name.

Signed-off-by: Charlie Egan <charlie@styra.com>