linter: Allow previously generated aggregate data
This extends the linter to allow previously generated aggregate data to
to be supplied to the linter. This is intended to be used in when
generating aggregate data for all files is costly and needs to be
cached. With this change, the cached data can be supplied to the linter
to allow it to be used without being regenerated.

Signed-off-by: Charlie Egan <charlie@styra.com>