linter: support single file aggregate data collection and parameterised aggregate data in Lint() (#1139)
* linter: allow populating aggregates w/ single file

This makes it possible to have the linter query for aggregate data even
when running with a single file.

Signed-off-by: Charlie Egan <charlie@styra.com>

* linter: Allow previously generated aggregate data

This extends the linter to allow previously generated aggregate data to
to be supplied to the linter. This is intended to be used in when
generating aggregate data for all files is costly and needs to be
cached. With this change, the cached data can be supplied to the linter
to allow it to be used without being regenerated.

Signed-off-by: Charlie Egan <charlie@styra.com>

* linter: Use WithAlwaysAggregate as new func name

Signed-off-by: Charlie Egan <charlie@styra.com>

---------

Signed-off-by: Charlie Egan <charlie@styra.com>