Adapt to new location format
Wow, this took quite some time to finish! But looking at the end
result, I'd say it was worth it. Having Roast represent locations
as `row:col:end_row:end_col` means quite a reduction in AST size,
and Regal's own JSON AST goes from 12 Mb to 10 Mb (in OPA AST that
is 24 Mb for comparison).

Running `regal lint bundle` is also ~100 ms faster, and improvement
of about 5%. Nice!

But this wasn't only about performance. Having end locations available
for any node means we can now easily show e.g. violations with a precise
location in LSP editors, and won't have to default to end-of-line, as
we previously did for most rules. This also meant that we were able to
remove all code in Regal that tried to calculate end position based
on the text attribute. Nice!

Signed-off-by: Anders Eknert <anders@styra.com>