Rule: `defer-assignment`
Finally a new rule in the `performance` category! And this is quite
a fun one. I was curious to see how many violations this would find
in the Regal codebase, if any. It did find a few, and good ones to
fix!

In order to keep it simple, this first implementation is perhaps a
bit too cautious about what it considers deferrable. Check the
implementation to see what exactly, but pretty much only assignments
where the next line is a not an assignment, a potential loop, or
something else that *might* make deferring the assignment undesirable.

Fixes #1147

Signed-off-by: Anders Eknert <anders@styra.com>