lsp: Fix inconsistent processing of ignores
In the LS linting, pattern/* would not work when it would in the CLI.
This was related to how abs paths were resolved and is fundamentally
caused by and issue with URIs being used as file names in the linter.

Now, all files from the LS lint are ignored based on their path with the
workspace root URI trimmed.

Signed-off-by: Charlie Egan <charlie@styra.com>