Fix custom rules that report on the absence of aggregates (#1260)
While custom aggregate rules would largely work before, those that report
on the absence of data (like a rule named "allow") did not work. This as
the aggregator sent nothing back in those caes — and contrary to the built-in
rules, Regal had then no way of knowing in the next step that the aggregate
report rules should be called even with no data.

I'm not super happy about the way we communicate this back to the Go side,
but there's a lot I'm not happy about with regards to how aggregate data
is shuffled back and forth. Some effort to fix that will however have to
wait until later.

Fixes #1259

Signed-off-by: Anders Eknert <anders@styra.com>