Enable a few more golangci-lint rules (#1280)
Some of these were broken in the past. Enabling those now and one
of them caught an actual issue, where we'd shadow the value of
`err` in the command worker, and may not have some errors reported
as we'd have expected.

The nolintlint rule that checks ignore directives may turn out to
be too annoying, but it was at least good to have it help clear out
comments that were not redundant. Leaving it on for now and we can
always disable if it turns out to be irritating.

Signed-off-by: Anders Eknert <anders@styra.com>