lsp: Rename correctly when conflicting (#1334)
* lsp: Rename correctly when conflicting

This updates the server's rename quick action to use the same name
generation conflict resolution logic as the fixer.

This is done by allowing the server's cache to be used as a new file
provider and consolidating the logic for generating changes required for
the templating worker and command worker.

This also changes the inlay hints to exit early if the file is not
parsed. This reduces errors in the server logs when working with empty
files.

Signed-off-by: Charlie Egan <charlie@styra.com>

* chore: Use strings for fixing file content

Signed-off-by: Charlie Egan <charlie@styra.com>

* lsp: pass version lookup to ToInput

Signed-off-by: Charlie Egan <charlie@styra.com>

* fixer: Add test to show fixes after rename

Signed-off-by: Charlie Egan <charlie@styra.com>

* rules: VersionsMap for all instances of input

This adapts the LSP and fixer to use the versionsMap functionality in
the same way as the linter.

Signed-off-by: Charlie Egan <charlie@styra.com>

---------

Signed-off-by: Charlie Egan <charlie@styra.com>