lsp: Use nested workspace configuration if found (#1372)
* lsp: Use nested workspace configuration if found

This makes the configuration for the server closer to the configuration
for the lint command. A config directory or file found within the
    client-supplied workspace root is used instead of only searching
    upwards.

If more than one .regal* root is found within the dir, then the first is
used. Ideally, we'd support many different config files but this is
harder to implement.

If none are found, then the previous behavior of searching upwards is
used.

Signed-off-by: Charlie Egan <charlie@styra.com>

* lsp: load new .regal.yaml files on create

This previously used the .regal dir only.

Signed-off-by: Charlie Egan <charlie@styra.com>

* lsp: Change config finding wording

Signed-off-by: Charlie Egan <charlie@styra.com>

---------

Signed-off-by: Charlie Egan <charlie@styra.com>