Performance improvements for `non-loop-expression`
Like we predicted, this rule costs a lot. 6 million allocations compared
to main is probably a record! But the value of the rule is also great, so
we'll have to make the best of it. Some perf tweaks help shave off at least
1 million allocations, and we're now down to 5.

Next up, I'll be taking a look at where most of the cost in OPA is while
evaluating this.

Funnily, I also found a missing case in the `use-object-keys` rule I added
just today, from seeing a set comprehension being used to get keys but not
getting flagged by the rule. Turned out `object.keys` was also noticeably
faster :)

```
1872039250 ns/op	3277562680 B/op	64585680 allocs/op // main
1990034458 ns/op	3577600880 B/op	70427856 allocs/op // non-loop-expression PR
1981020667 ns/op	3515957384 B/op	69337090 allocs/op // use built-ins to get _exprs row
1976604792 ns/op	3518227272 B/op	69422952 allocs/op // inline ast.is_wildcard
1957780375 ns/op	3518962424 B/op	69226825 allocs/op // uglier loops
1923554667 ns/op	3501324176 B/op	68954057 allocs/op // object.keys(sps)
```

Signed-off-by: Anders Eknert <anders@styra.com>