lsp: URI/Path handling fixes (#1419)
There was an issue where a colon in a file name would cause the wrong
path to be generated from a URI and vice versa. This caused the server
to reload the file every second as it was missing, but really it was
just registered under the wrong URI in the cache.

Signed-off-by: Charlie Egan <charlie@styra.com>