lsp: Fix Code Lenses jank
Previously, when editing a line, if parse errors come and go as you
edit, the Code Lenses will toggle on and off as the parse error list is
updated.

Now, only if the line count is different from the last successful parse,
will the Code Lenses be disabled.

This means that a line can be typed while it contains errors, and code
lenses will now come and go.

Signed-off-by: Charlie Egan <charlie@styra.com>