Parameterize some tests (#1460)
This was mostly done to try things out last night. Not sure yet if there's
one particular model / style for how we should do this, but some alternative
approaches have been tested here. Commiting this in the meantime for us to
have some real examples to keep that discussion around.

As a bonus, the number of tests reported are now 701, compared to 685 without
parameterization :)

Signed-off-by: Anders Eknert <anders@styra.com>