Add max-test-rule-length setting to rule-length rule (#1476)
Since tests are often mostly data, there is little to gain
by moving relevant data outside of tests. This PR introduces
a max-test-rule-length setting, which by default is set to 60
rather than 30, which is the default for other rules.

Finally getting rid of all inline ignore directives related
to this in our tests!

Fixes #1125

Signed-off-by: Anders Eknert <anders@styra.com>