Rule: narrow-argument
This was a lot of fun! Check out the docs included in this PR for more
information about this rule and potential caveats. But all things
considered, I think this rule is great and have enabled it for our own
use (with a few exceptions configured).

Even when narrowing isn't the best option, I think it's great to be
forced to think about the dependencies of functions as I work on them.

Consistently named args help a lot here, as being able to exclude commonly
passed values, like `cfg` in our case, "fixed" almost all the reported
violations, and I only had to use a handful of ignore directives for the
rest. But most of the issues reported were rewritten, and I think for the
better!

(a few things changed here were not directly reported, but things I saw
could be improved in the files I opened)

Fixes #1480

Signed-off-by: Anders Eknert <anders@styra.com>