perf: Improve how config is parsed, and when
**main**
```
1188808042 ns/op	3332404200 B/op	63908530 allocs/op
```
**this**
```
1148661583 ns/op	3173734464 B/op	61290539 allocs/op
```

2,7 million allocs!

Mostly guided by `--profile` metrics, which suggested a lot of
time being spent in config.rego. That is interesting, as there
isn't really anything there that would easily map to the pprof
heap dumps. And there weren't any real bottlenecks to be found
either. Just custom functions called very frequently, and some
objects getting unioned constantly, even though the info added
(the calculated level) almost never matters, as the main route
won't even consider ignored policies.

This makes it much cheaper to check for the actual level for a
rule from the main.rego composition, while also making it much
cheaper to get the config for a rule by *not* merging the data
(level) we normally don't care for. So a win-win really!

Signed-off-by: Anders Eknert <anders@styra.com>