Scan nested expressions
A few of the first Regal rules only ever checked the top level of
a body, and would not traverse into e.g. comprehensions or every
bodies. This has been annoying for long, and while doing this
properly is certainly more expensive (+2 million allocs) I think
we can afford that at this point in order to prioritize correctness.
May very well be that many will have Regal find "new" issues in the
next version due to this.

Also:
- We now also separate ast.found.references and ast.found.calls. This
  too comes at a cost, but makes it much more obvious what our code is
  doing, and we can now afforc prioritizing that too.
- Improve `narrow-argument` rule to also take array indices into account
  for narrowing (i.e. only `arr[1].value[0]` referenced)
- Fix new golangci-lint rule that says public functions should be above
  private ones.

Fixes #82

Signed-off-by: Anders Eknert <anders@styra.com>